ZigZagBoomerang

Back to repository: [https://github.com/mschauer/ZigZagBoomerang.jl]

ZigZagBoomerang.Boomerang

ZigZagBoomerang.ContinuousDynamics

ZigZagBoomerang.LocalZigZag

ZigZagBoomerang.ab

ZigZagBoomerang.discretization

ZigZagBoomerang.move_forward

ZigZagBoomerang.move_forward

ZigZagBoomerang.move_forward!

ZigZagBoomerang.neighbours

ZigZagBoomerang.pdmp

ZigZagBoomerang.pdmp

ZigZagBoomerang.pdmp_inner!

ZigZagBoomerang.poisson_time

ZigZagBoomerang.poisson_time

ZigZagBoomerang.reflect!

ZigZagBoomerang.λ

ZigZagBoomerang.λ_bar

ZigZagBoomerang.Boomerang — Type

Boomerang(μ, λ) <: ContinuousDynamics

Dynamics preserving the N(μ, 1) measure (Boomerang) with refreshment time λ

ZigZagBoomerang.ContinuousDynamics — Type

ContinuousDynamics

Abstract type for the deterministic dynamics of PDMPs

ZigZagBoomerang.LocalZigZag — Type

struct LocalZigZag <: ContinuousDynamics

Flag for local implementation of the ZigZag which exploits any conditional independence structure of the target measure

ZigZagBoomerang.ab — Method

ab(G, i, x, θ, c, Z::LocalZigZag)

Returns the constant term a and linear term b when computing the Poisson times from the upper upper bounding rates λᵢ(t) = max(a + b*t)^2. The factors a and b can be function of the current position x, velocity θ, tuning parameter c and the Graph G

ZigZagBoomerang.discretization — Method

discretization(x::Vector{Skeleton}, Flow::Boomerang, dt)

Tansform the output of the algorithm (a skeleton of points) to a trajectory.

ZigZagBoomerang.move_forward! — Method

move_forward!(τ, t, x, θ, Z::LocalZigZag)

Updates the position x, velocity θ and time t of the process after a time step equal to τ according to the deterministic dynamics of the LocalZigZag sampler: (x(τ), θ(τ)) = (x(0) + θ(0)*t, θ(0)). x: current location, θ: current velocity, t: current time,

ZigZagBoomerang.move_forward — Method

move_forward(τ, t, x, θ, B::Boomerang)

Updates the position x, velocity θ and time t of the process after a time step equal to τ according to the deterministic dynamics of the Boomerang sampler: xt = μ +(x0 − μ)cos(t) + v_0sin(t), vt = −(x0 − μ)sin(t) + v_0cos(t) x: current location, θ: current velocity, t: current time.

ZigZagBoomerang.move_forward — Method

move_forward(τ, t, x, θ, ::ZigZag)

Updates the position x, velocity θ and time t of the process after a time step equal to τ according to the deterministic dynamics of the ZigZag sampler: (x(τ), θ(τ)) = (x(0) + θ(0)*t, θ(0)). x: current location, θ: current velocity, t: current time,

ZigZagBoomerang.neighbours — Method

neighbours(G::Vector{<:Pair}, i) = G[i].second

Return extended neighbourhood of i including i. G: graphs of neightbourhoods

ZigZagBoomerang.pdmp — Method

pdmp(G, ∇ϕ, t0, x0, θ0, T, c, Z::LocalZigZag; factor=1.5, adapt=false)

LocalZigZag algorithm. Input: Graph G, gradient of negative log density ∇ϕ, initial time t0, initial position x0, initial velocity θ0, final clock T, tuning parameter c.

The process moves at to time T with invariant mesure μ(dx) ∝ exp(-ϕ(x))dx and outputs a collection of reflection points Ξ which, together with the initial triple x θ and t are sufficient for reconstructuing continuously the continuous path

ZigZagBoomerang.pdmp — Method

pdmp(∇ϕ, x, θ, T, Flow::ContinuousDynamics; adapt=true,  factor=2.0)

Run a piecewise deterministic process from location and velocity x, θ until time T. c is a tuning parameter for the upper bound of the Poisson rate. If adapt = false, c = c*factor is tried, otherwise an error is thrown.

Returns vector of tuples (t, x, θ) (time, location, velocity) of direction change events.

ZigZagBoomerang.pdmp_inner! — Method

pdmp_inner!(Ξ, G, ∇ϕ, x, θ, Q, t, c, (num, acc), Z::LocalZigZag;
    factor=1.5, adapt=false)

Inner loop of the LocalZigZag algorithm. Input: a dependency graph G, gradient ∇ϕ, current position x, velocity θ, Queue of events Q, time t, and tuning parameter c.

The sampler 1) extracts from the queue the first event time. 2) moves deterministically according to the LocalZigZag dynamics until event time. 3) Evaluates whether the event time is a reflection time or not. 4) If it is a reflection time, the velocity reflects according the the LocalZigZag reflection rule and updates Q according to the dependency graph G. (num, acc) counts how many event times occour and how many of those are real reflection times.

ZigZagBoomerang.poisson_time — Method

poisson_time(a, b, u)

Obtaining waiting time for inhomogeneous Poisson Process with rate of the form λ(t) = (a + b*t)^+, a,b ∈ R, u uniform random variable

ZigZagBoomerang.poisson_time — Method

poisson_time(a, u)

Obtaining waiting time for homogeneous Poisson Process with rate of the form λ(t) = a, a ≥ 0, u uniform random variable

ZigZagBoomerang.reflect! — Method

    reflect!(i, θ, x, Z)

Reflection rule of LocalZigZag sampler at reflection time. i: coordinate which flips sign, θ: velocity, x: position (not used for the LocalZigZag)

ZigZagBoomerang.λ — Method

λ(∇ϕ, i, x, θ, Z::LocalZigZag)

ith Poisson rate of the LocalZigZag sampler

ZigZagBoomerang.λ_bar — Method

λ_bar(G, i, x, θ, c, Z::LocalZigZag)

Computes the bounding rate λ_bar at position x and velocity θ.

Literature

Joris Bierkens, Paul Fearnhead, Gareth Roberts: The Zig-Zag Process and Super-Efficient Sampling for Bayesian Analysis of Big Data. The Annals of Statistics, 2019, 47. Vol., Nr. 3, pp. 1288-1320. [https://arxiv.org/abs/1607.03188].

Joris Bierkens, Sebastiano Grazzi, Kengo Kamatani and Gareth Robers: The Boomerang Sampler. ICML 2020. [https://arxiv.org/abs/2006.13777].

Joris Bierkens, Sebastiano Grazzi, Frank van der Meulen, Moritz Schauer: A piecewise deterministic Monte Carlo method for diffusion bridges. 2020. [https://arxiv.org/abs/2001.05889].

https://github.com/jbierkens/ICML-boomerang/