ZigZagBoomerang

Back to repository: https://github.com/mschauer/ZigZagBoomerang.jl

ZigZagBoomerang.Boomerang

ZigZagBoomerang.Boomerang1d

ZigZagBoomerang.Bps

ZigZagBoomerang.ContinuousDynamics

ZigZagBoomerang.FactBoomerang

ZigZagBoomerang.FactTrace

ZigZagBoomerang.ZigZag

ZigZagBoomerang.ZigZag1d

ZigZagBoomerang.Trace

ZigZagBoomerang.ab

ZigZagBoomerang.discretize

ZigZagBoomerang.discretize

ZigZagBoomerang.idot

ZigZagBoomerang.move_forward

ZigZagBoomerang.move_forward

ZigZagBoomerang.move_forward!

ZigZagBoomerang.move_forward!

ZigZagBoomerang.neighbours

ZigZagBoomerang.normsq

ZigZagBoomerang.pdmp

ZigZagBoomerang.pdmp

ZigZagBoomerang.pdmp_inner!

ZigZagBoomerang.poisson_time

ZigZagBoomerang.poisson_time

ZigZagBoomerang.pos

ZigZagBoomerang.reflect!

ZigZagBoomerang.spdmp

ZigZagBoomerang.λ

ZigZagBoomerang.λ

ZigZagBoomerang.λ_bar

ZigZagBoomerang.Boomerang — Type

Boomerang(μ, λ) <: ContinuousDynamics

Dynamics preserving the N(μ, I) measure (Boomerang) with refreshment time λ

ZigZagBoomerang.Boomerang1d — Type

Boomerang1d(Σ, μ, λ) <: ContinuousDynamics

1-d toy boomerang samper. Dynamics preserving the N(μ, Σ) measure with refreshment time λ.

ZigZagBoomerang.Bps — Type

Bps(λ) <: ContinuousDynamics

Bouncy particle sampler, λ is the refreshment rate, which has to be strictly positive.

ZigZagBoomerang.ContinuousDynamics — Type

ContinuousDynamics

Abstract type for the deterministic dynamics of PDMPs

ZigZagBoomerang.FactBoomerang — Type

FactBoomerang(Γ, μ, λ) <: ContinuousDynamics

Factorized Boomerang dynamics preserving the N(μ, inv(Diagonal(Γ))) measure with refreshment time λ.

Exploits the conditional independence structure of the target measure, in form the argument Γ, a sparse precision matrix approximating target precision. μ is the approximate target mean.

ZigZagBoomerang.FactTrace — Type

FactTrace

See Trace.

ZigZagBoomerang.ZigZag — Type

struct ZigZag(Γ, μ) <: ContinuousDynamics

Local ZigZag sampler which exploits any independence structure of the target measure, in form the argument Γ, a sparse precision matrix approximating target precision. μ is the approximate target mean.

ZigZagBoomerang.ZigZag1d — Type

ZigZag1d <: ContinuousDynamics

1-d toy ZigZag sampler, dynamics preserving the Lebesgue measure.

ZigZagBoomerang.Trace — Method

Trace(t0::T, x0, θ0, F::Union{ZigZag,FactBoomerang})

Trace object for exact trajectory of pdmp samplers. Returns an iterable FactTrace object. Note that iteration iterates pairs t => x where the vector x is modified inplace, so copies have to be made if the x is to be saved. collect applied to a trace object automatically copies x. discretize returns a discretized version.

ZigZagBoomerang.ab — Method

ab(G, i, x, θ, c, Flow)

Returns the constant term a and linear term b when computing the Poisson times from the upper upper bounding rates λᵢ(t) = max(a + b*t)^2. The factors a and b can be function of the current position x, velocity θ, tuning parameter c and the Graph G

ZigZagBoomerang.discretize — Method

discretize(trace::FactTrace, dt)

Discretize trace with step-size dt. Returns iterable object iterating pairs t => x.

Iteration changes the vector x inplace, collect creates copies.

ZigZagBoomerang.discretize — Method

discretize(x::Vector, Flow, dt)

Transform the output of the algorithm (a skeleton of points) to a trajectory. Simple 1-d version.

ZigZagBoomerang.idot — Method

idot(A, j, x) = dot(A[:, j], x)

Compute column-vector dot product exploiting sparsity of A.

ZigZagBoomerang.move_forward! — Method

move_forward!(τ, t, x, θ, B::Boomerang)

Updates the position x, velocity θ and time t of the process after a time step equal to τ according to the deterministic dynamics of the Boomerang sampler which are the Hamiltonian dynamics preserving the Gaussian measure: : xt = μ +(x0 − μ)cos(t) + v_0sin(t), vt = −(x0 − μ)sin(t) + v_0cos(t) x: current location, θ: current velocity, t: current time.

ZigZagBoomerang.move_forward! — Method

move_forward!(τ, t, x, θ, Z::Union{Bps, ZigZag})

Updates the position x, velocity θ and time t of the process after a time step equal to τ according to the deterministic dynamics of the Buoncy particle sampler (Bps) and ZigZag: (x(τ), θ(τ)) = (x(0) + θ(0)*t, θ(0)). x: current location, θ: current velocity, t: current time,

ZigZagBoomerang.move_forward — Method

move_forward(τ, t, x, θ, B::Boomerang1d)

Updates the position x, velocity θ and time t of the process after a time step equal to τ according to the deterministic dynamics of the Boomerang1d sampler: xt = μ +(x0 − μ)cos(t) + v_0sin(t), vt = −(x0 − μ)sin(t) + v_0cos(t) x: current location, θ: current velocity, t: current time.

ZigZagBoomerang.move_forward — Method

move_forward(τ, t, x, θ, ::ZigZag1d)

Updates the position x, velocity θ and time t of the process after a time step equal to τ according to the deterministic dynamics of the ZigZag1d sampler: (x(τ), θ(τ)) = (x(0) + θ(0)*t, θ(0)). x: current location, θ: current velocity, t: current time,

ZigZagBoomerang.neighbours — Method

neighbours(G::Vector{<:Pair}, i) = G[i].second

Return extended neighbourhood of i including i. G: graphs of neightbourhoods

ZigZagBoomerang.normsq — Method

normsq(x)

Squared 2-norm.

ZigZagBoomerang.pdmp — Method

pdmp(∇ϕ, t0, x0, θ0, T, c, F::Union{ZigZag, FactBoomerang}, args...; factor=1.5, adapt=false) = Ξ, (t, x, θ), (acc, num)

Outer loop of the factorised samplers, the Factorised Boomerang algorithm and the Zig-Zag sampler. Inputs are a function ∇ϕ giving ith element of gradient of negative log target density ∇ϕ(x, i, args...), starting time and position t0, x0, velocities θ0, and tuning vector c for rejection bounds and final clock T.

The process moves to time T with invariant mesure μ(dx) ∝ exp(-ϕ(x))dx and outputs a collection of reflection points which, together with the initial triple t, x θ are sufficient for reconstructuing continuously the continuous path. It returns a FactTrace (see Trace) object Ξ, which can be collected into pairs t => x of times and locations and discretized with discretize. Also returns the number of total and accepted Poisson events and updated bounds c (in case of adapt==true the bounds are multiplied by factor if they turn out to be too small.)

This version does not assume that ∇ϕ has sparse conditional dependencies, see spdmp.

ZigZagBoomerang.pdmp — Method

pdmp(∇ϕ, x, θ, T, Flow::ContinuousDynamics; adapt=true,  factor=2.0)

Run a piecewise deterministic process from location and velocity x, θ until time T. c is a tuning parameter for the upper bound of the Poisson rate. If adapt = false, c = c*factor is tried, otherwise an error is thrown.

Returns vector of tuples (t, x, θ) (time, location, velocity) of direction change events.

ZigZagBoomerang.pdmp_inner! — Method

pdmp_inner!(Ξ, G, ∇ϕ, t, x, θ, Q, c, (acc, num),
    F::Union{ZigZag,FactBoomerang}, args...; factor=1.5, adapt=false)
    = t, x, θ, (acc, num), c

Inner loop of the factorised samplers: the factorised Boomerang algorithm and the Zig-Zag sampler. Given a dependency graph G, gradient ∇ϕ, current position x, velocity θ, Queue of events Q, time t, and tuning parameter c.

The sampler 1) extracts from the queue the first event time. 2) moves deterministically according to its dynamics until event time. 3) Evaluates whether the event time is a accepted reflection time or not. 4) If it is a reflection time, the velocity reflects according its reflection rule and updates Q according to the dependency graph G. The sampler proceeds time until the next accepted reflection time or refreshment time. (num, acc) incrementally counts how many event times occour and how many of those are real reflection times.

ZigZagBoomerang.poisson_time — Method

poisson_time(a, b, u)

Obtaining waiting time for inhomogeneous Poisson Process with rate of the form λ(t) = (a + b*t)^+, a,b ∈ R, u uniform random variable

ZigZagBoomerang.poisson_time — Method

poisson_time(a[, u])

Obtaining waiting time for homogeneous Poisson Process with rate of the form λ(t) = a, a ≥ 0, u uniform random variable

ZigZagBoomerang.pos — Method

pos(x)

Positive part of x (i.e. max(0,x)).

ZigZagBoomerang.reflect! — Method

    reflect!(i, x, θ, F)

Reflection rule of sampler F at reflection time. i: coordinate which flips sign, x: position, θ: velocity (position not used for the ZigZag and FactBoomerang.)

ZigZagBoomerang.spdmp — Method

spdmp(∇ϕ, t0, x0, θ0, T, c, F::Union{ZigZag,FactBoomerang}, args...;
    factor=1.5, adapt=false)

Version of spdmp which assumes that i only depends on coordinates x[j] for j in neighbours(G, i).

ZigZagBoomerang.λ — Method

λ(∇ϕ, i, x, θ, Z::FactBoomerang)

ith Poisson rate of the FactBoomerang sampler

ZigZagBoomerang.λ — Method

λ(∇ϕ, i, x, θ, Z::ZigZag)

ith Poisson rate of the ZigZag sampler

ZigZagBoomerang.λ_bar — Method

λ_bar(G, i, x, θ, c, Z)

Computes the bounding rate λ_bar at position x and velocity θ.

Literature

Joris Bierkens, Paul Fearnhead, Gareth Roberts: The Zig-Zag Process and Super-Efficient Sampling for Bayesian Analysis of Big Data. The Annals of Statistics, 2019, 47. Vol., Nr. 3, pp. 1288-1320. https://arxiv.org/abs/1607.03188.

Joris Bierkens, Sebastiano Grazzi, Kengo Kamatani and Gareth Robers: The Boomerang Sampler. ICML 2020. https://arxiv.org/abs/2006.13777.

Joris Bierkens, Sebastiano Grazzi, Frank van der Meulen, Moritz Schauer: A piecewise deterministic Monte Carlo method for diffusion bridges. 2020. https://arxiv.org/abs/2001.05889.

https://github.com/jbierkens/ICML-boomerang/ (code accompanying the paper "The Boomerang Sampler")