ZigZagBoomerang

Back to repository: [https://github.com/mschauer/ZigZagBoomerang.jl]

ZigZagBoomerang.Boomerang

ZigZagBoomerang.ContinuousDynamics

ZigZagBoomerang.discretization

ZigZagBoomerang.pdmp

ZigZagBoomerang.poisson_time

ZigZagBoomerang.poisson_time

ZigZagBoomerang.Boomerang — Type

Boomerang(λref) <: ContinuousDynamics

Dynamics preserving the standard Gaussian measure (Boomerang) with refreshment time λref

ZigZagBoomerang.ContinuousDynamics — Type

ContinuousDynamics

Abstract type for the deterministic dynamics of PDMPs

ZigZagBoomerang.discretization — Method

discretization(x::Vector{Skeleton}, Flow::Boomerang, dt)

Tansform the output of the algorithm (a skeleton of points) to a trajectory.

ZigZagBoomerang.pdmp — Method

pdmp(∇ϕ, x, θ, T, Flow::ContinuousDynamics; adapt=true,  factor=2.0)

Run a piecewise deterministic process from location and velocity x, θ until time T. c is a tuning parameter for the upper bound of the Poisson rate. If adapt = false, c = c*factor is tried, otherwise an error is thrown.

Returns vector of tuples (t, x, θ) (time, location, velocity) of direction change events.

ZigZagBoomerang.poisson_time — Method

poisson_time(a, b, u)

obtaining waiting time for inhomogeneous Poisson Process with rate of the form λ(t) = (a + b*t)^+, a,b ∈ R, u uniform random variable

ZigZagBoomerang.poisson_time — Method

h_poisson_time(a, u)

obtaining waiting time for homogeneous Poisson Process with rate of the form λ(t) = a, a ≥ 0, u uniform random variable

Literature

Joris Bierkens, Paul Fearnhead, Gareth Roberts: The Zig-Zag Process and Super-Efficient Sampling for Bayesian Analysis of Big Data. The Annals of Statistics, 2019, 47. Vol., Nr. 3, pp. 1288-1320. [https://arxiv.org/abs/1607.03188].

Joris Bierkens, Sebastiano Grazzi, Kengo Kamatani and Gareth Robers: The Boomerang Sampler. ICML 2020. [https://arxiv.org/abs/2006.13777].

Joris Bierkens, Sebastiano Grazzi, Frank van der Meulen, Moritz Schauer: A piecewise deterministic Monte Carlo method for diffusion bridges. 2020. [https://arxiv.org/abs/2001.05889].

https://github.com/jbierkens/ICML-boomerang/